fix: Address all P1 and code-quality findings from PR #122 bot reviews
- P1 (Codex): Include work item index in attachment ZIP paths to prevent silent drops when two emails share attachment filenames
- P1 (droid): Reduce --count upper bound to int.MaxValue-1 to prevent DAT header + ChaosEngine overflow
- ConcordanceWriter: Use request.ColumnDelimiter (ASCII 20) instead of hardcoded comma, per REQ_E-017/REQ-049
- ChaosEngine: Remove unnecessary (long)j cast
- OptWriter: Remove dead Random plumbing
- Tests: Strengthen Concordance escaping assertion, add count-exceeding-max test, update delimiter tests
- Docs: Update README count bound

+++ src/ChaosEngine.cs
@@ -208,7 +208,7 @@ internal class ChaosEngine
             long candidate = random.NextInt64(1, j + 1);
             if (!selected.Add(candidate))
             {
-                selected.Add((long)j);
+                selected.Add(j);
             }
         }
